Backend/requirements.txt
asgiref==3.9.1
bleach==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==65.11.2
django-stubs==5.2.5
django-stubs-ext==5.2.5
djangorestframework==3.16.1
djangorestframework-stubs==3.16.2
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
idna==3.10
oauthlib==3.3.1
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
urllib3==2.5.0
webencodings==0.5.1
